src/main/java/uk/gov/hmcts/divorce/divorcecase/model/access/CaseworkerAndSuperUserAccess.java
package uk.gov.hmcts.divorce.divorcecase.model.access;

import com.google.common.collect.HashMultimap;
import com.google.common.collect.SetMultimap;
import uk.gov.hmcts.ccd.sdk.api.HasAccessControl;
import uk.gov.hmcts.ccd.sdk.api.HasRole;
import uk.gov.hmcts.ccd.sdk.api.Permission;

import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CASEWORKER_COURTADMIN;
import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CASEWORKER_LEGAL_ADVISOR;
import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CASEWORKER_SUPERUSER;
import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.SOLICITOR;

public class CaseworkerAndSuperUserAccess implements HasAccessControl {

    @Override
    public SetMultimap<HasRole, Permission> getGrants() {
        SetMultimap<HasRole, Permission> grants = HashMultimap.create();
        grants.putAll(SOLICITOR, Permissions.READ);
        grants.putAll(CASEWORKER_LEGAL_ADVISOR, Permissions.READ);

        grants.putAll(CASEWORKER_SUPERUSER, Permissions.CREATE_READ_UPDATE_DELETE);

        grants.putAll(CASEWORKER_COURTADMIN, Permissions.CREATE_READ_UPDATE);

        return grants;
    }
}
